% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/featureReport.R
\name{featureReport}
\alias{featureReport}
\title{Create a Features Report of a Feature in the Data.frame}
\usage{
featureReport(vector, vecName, ...)
}
\arguments{
\item{vector}{A vector.}

\item{vecName}{The Name of the vector.}
}
\description{
featureReport will create a list of information of the feature
}
\examples{
featureReport(iris[,1])
}

