% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitrFeature.R
\name{knitrFeature}
\alias{knitrFeature}
\title{Function for Knitring Feature Report Function}
\usage{
knitrFeature(feaReport)
}
\arguments{
\item{feaReport}{featureReport created by featureReport function.}
}
\description{
Function for Knitring Feature Report Function
}
\examples{
feaReport = featureReport(iris[,1])
knitrFeature(feaReport)
}

