% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitrNa.R
\name{knitrNa}
\alias{knitrNa}
\title{Function for Knitring NA Report Function}
\usage{
knitrNa(naR)
}
\arguments{
\item{naR}{naReport created by naReport function.}
}
\description{
Function for Knitring NA Report Function
}
\examples{
report = naReport(testData)
knitrNa(report)
}

