% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitrUniq.R
\name{knitrUniq}
\alias{knitrUniq}
\title{Function for Knitring Unique Report Function}
\usage{
knitrUniq(uniqR)
}
\arguments{
\item{uniqR}{uniqReport created by uniqueReport function.}
}
\description{
Function for Knitring Unique Report Function
}
\examples{
report = uniqueReport(testData)
knitrUniq(report)
}

