% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naReport.R
\name{naReport}
\alias{naReport}
\title{Create a naReport of a data.frame}
\usage{
naReport(inputData,range = c(0, 0.4))
}
\arguments{
\item{inputData}{The data frame to create report before modeling.}

\item{range}{A vector of length 2 for the range of NA proportion.
Variables with NA proportion less than range[1] will do nothing,
Variables with NA proportion more than range[2] will be suggested to delete.
Others will be suggested to remove the observations.}
}
\description{
naReport will detect the NAs i around the variables in the
dataframe and show the codes users may need for omit.
}
\examples{
naReport(testData)
}

