% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preReport.R
\name{preReport}
\alias{preReport}
\title{Create a preReport of a data.frame}
\usage{
preReport(iris)

preReport(inputData,
         path,
         scaffold = system.file("skeleton", package = "preReport"),
         open_rmd = T,
         useBytes = T, ...)
}
\arguments{
\item{inputData}{The data frame to create report before modeling.}

\item{path}{A vector of length 2 for the range of NA proportion.}

\item{scaffold}{Path for the file used in the html file.}

\item{open_rmd}{Whether open the Rmd file or not.}

\item{useBytes}{If useByte when writing the Rmd file.}

\item{...}{Other parameters put into writeLines function.}
}
\description{
preReport will create a Rmd Report from a data.frame.
}
\examples{
preReport(iris)
}

