% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uniqueReport.R
\name{uniqueReport}
\alias{uniqueReport}
\title{Create a uniqueReport of a data.frame}
\usage{
uniqueReport (inputData,range = c(0, 0.4))
}
\arguments{
\item{inputData}{The data frame to create report before modeling.}

\item{range}{A vector of length 2 for the range of NA proportion.
Variables with NA proportion less than range[1] will do nothing,
Variables with NA proportion more than range[2] will be suggested to delete.
Others will be suggested to remove the observations.}
}
\description{
uniqueReport will create a unique Report from a data.frame.
}
\examples{
uniqueReport(testData)
}

